requirements.txt
absl-py==2.3.1
altair==5.5.0
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1749054771823/work
annotated-types @ file:///work/perseverance-python-buildout/croot/annotated-types_1728386958747/work
archspec @ file:///croot/archspec_1709217642129/work
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
boltons @ file:///croot/boltons_1751383775419/work
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
cachetools==6.2.1
certifi @ file:///croot/certifi_1752653701669/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.3.0
conda @ file:///croot/conda_1749171252582/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1751061994888/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1752801564376/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1728487058234/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1751307506088/work
cryptography @ file:///croot/cryptography_1748977429425/work
distro @ file:///work/perseverance-python-buildout/croot/distro_1728396110052/work
flatbuffers==25.9.23
frozendict @ file:///work/perseverance-python-buildout/croot/frozendict_1728497542215/work
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna @ file:///work/perseverance-python-buildout/croot/idna_1728385935861/work
Jinja2==3.1.6
jsonpatch @ file:///work/perseverance-python-buildout/croot/jsonpatch_1728399595941/work
jsonpointer==2.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
libclang==18.1.1
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
Markdown==3.10
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1728387994477/work
MarkupSafe==3.0.3
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1728387286143/work
menuinst @ file:///croot/menuinst_1750861706497/work
ml_dtypes==0.5.3
namex==0.1.0
narwhals==2.11.0
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.3.3
pillow==12.0.0
platformdirs @ file:///croot/platformdirs_1744273042065/work
pluggy @ file:///croot/pluggy_1733169602837/work
protobuf==6.33.0
pyarrow==21.0.0
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1750768261732/work
pydantic_core @ file:///croot/pydantic-core_1750754736071/work
pydeck==0.9.1
Pygments @ file:///croot/pygments_1744664109463/work
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1728386193338/work
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests @ file:///croot/requests_1750426329888/work
rich @ file:///croot/rich_1732638981168/work
rpds-py==0.28.0
ruamel.yaml @ file:///croot/ruamel.yaml_1745960305322/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1745937152469/work
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
toml==0.10.2
tornado==6.5.2
tqdm @ file:///croot/tqdm_1738943501192/work
truststore @ file:///croot/truststore_1736550121485/work
typing-inspection @ file:///croot/typing-inspection_1746023470701/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1750775463400/work
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
zstandard @ file:///croot/zstandard_1731356346222/work
